+++ packages/plugin-electron-client-state-persistence/src/signal_handler.h
@@ -1,7 +1,0 @@
-#pragma once
-
-void becsp_signal_install(void (*func)(int));
-
-void becsp_signal_uninstall(void);
-
-void becsp_signal_raise(int signal);