+++ packages/plugin-react-native-navigation/react-native-navigation.js
@@ -27,3 +27,6 @@ module.exports = class BugsnagPluginReactNativeNavigation {
     })
   }
 }
+
+// add a default export for ESM modules without interop
+module.exports.default = module.exports